+++ packages/protocol-devtools-solana/src/endpointv2/sdk.ts
@@ -1,4 +1,5 @@
 import { MessagingFee } from '@layerzerolabs/protocol-devtools'
+import { oft } from '@layerzerolabs/oft-v2-solana-sdk'
 import type {
     IEndpointV2,
     IUlnRead,
@@ -34,10 +35,12 @@ import {
     SimpleMessageLibProgram,
     UlnProgram,
 } from '@layerzerolabs/lz-solana-sdk-v2'
-import { Connection, PublicKey, Transaction } from '@solana/web3.js'
+import { Connection, PublicKey, Transaction, TransactionInstruction } from '@solana/web3.js'
 import assert from 'assert'
 import { Uln302 } from '@/uln302'
 import { SetConfigSchema } from './schema'
+import { createNoopSigner, publicKey, TransactionBuilder, WrappedInstruction } from '@metaplex-foundation/umi'
+import { fromWeb3JsPublicKey } from '@metaplex-foundation/umi-web3js-adapters'
 
 /**
  * Solana-specific SDK for EndpointV2 contracts
@@ -71,6 +74,15 @@ export class EndpointV2 extends OmniSDK implements IEndpointV2 {
         } catch (error) {
             throw new Error(`Failed to get delegate for ${this.label} for oapp ${oapp}: ${error}`)
         }
+    }
+
+    // Intentionally not marked as AsyncRetriable as it calls getDelegate(...) which is already AsyncRetriable.
+    protected async safeGetDelegate(oapp: OmniAddress): Promise<PublicKey> {
+        const delegate = await this.getDelegate(oapp)
+        if (delegate == null) {
+            throw new Error(`No delegate set for OApp ${oapp}`)
+        }
+        return new PublicKey(delegate)
     }
 
     async isDelegate(oapp: OmniAddress, delegate: OmniAddress): Promise<boolean> {
@@ -224,18 +236,44 @@ export class EndpointV2 extends OmniSDK implements IEndpointV2 {
 
         this.logger.debug(`Setting send library for eid ${eid} (${eidLabel}) and OApp ${oapp} to ${uln}`)
 
-        const instruction = await mapError(
-            () => this.program.setSendLibrary(this.userAccount, new PublicKey(oapp), new PublicKey(uln), eid),
-            (error) =>
-                new Error(`Failed to set the send library for ${this.label} and OApp ${oapp} for ${eidLabel}: ${error}`)
-        )
-
-        const transaction = new Transaction().add(instruction)
-
         return {
-            ...(await this.createTransaction(transaction)),
+            ...(await this.createTransaction(
+                this._umiToWeb3Tx([
+                    oft.setSendLibrary(
+                        {
+                            admin: createNoopSigner(fromWeb3JsPublicKey(await this.safeGetDelegate(oapp))),
+                            oftStore: publicKey(oapp),
+                        },
+                        {
+                            sendLibraryProgram: publicKey(uln),
+                            remoteEid: eid,
+                        }
+                    ),
+                ])
+            )),
             description: `Setting send library for eid ${eid} (${eidLabel}) and OApp ${oapp} to ${uln}`,
         }
+    }
+
+    // Convert Umi instructions to Web3JS Transaction
+    protected _umiToWeb3Tx(ixs: WrappedInstruction[]): Transaction {
+        const web3Transaction = new Transaction()
+        const txBuilder = new TransactionBuilder(ixs)
+        txBuilder.getInstructions().forEach((umiInstruction) => {
+            const web3Instruction = new TransactionInstruction({
+                programId: new PublicKey(umiInstruction.programId),
+                keys: umiInstruction.keys.map((key) => ({
+                    pubkey: new PublicKey(key.pubkey),
+                    isSigner: key.isSigner,
+                    isWritable: key.isWritable,
+                })),
+                data: Buffer.from(umiInstruction.data),
+            })
+
+            // Add the instruction to the Web3.js transaction
+            web3Transaction.add(web3Instruction)
+        })
+        return web3Transaction
     }
 
     async setReceiveLibrary(
@@ -256,25 +294,22 @@ export class EndpointV2 extends OmniSDK implements IEndpointV2 {
 
         this.logger.debug(`Setting receive library for eid ${eid} (${eidLabel}) and OApp ${oapp} to ${uln}`)
 
-        const instruction = await mapError(
-            () =>
-                this.program.setReceiveLibrary(
-                    this.userAccount,
-                    new PublicKey(oapp),
-                    new PublicKey(uln),
-                    eid,
-                    Number(gracePeriod)
-                ),
-            (error) =>
-                new Error(
-                    `Failed to set the receive library for ${this.label} and OApp ${oapp} for ${eidLabel}: ${error}`
-                )
-        )
-
-        const transaction = new Transaction().add(instruction)
-
         return {
-            ...(await this.createTransaction(transaction)),
+            ...(await this.createTransaction(
+                this._umiToWeb3Tx([
+                    oft.setReceiveLibrary(
+                        {
+                            admin: createNoopSigner(fromWeb3JsPublicKey(await this.safeGetDelegate(oapp))),
+                            oftStore: publicKey(oapp),
+                        },
+                        {
+                            receiveLibraryProgram: publicKey(uln),
+                            remoteEid: eid,
+                            gracePeriod,
+                        }
+                    ),
+                ])
+            )),
             description: `Setting receive library for eid ${eid} (${eidLabel}) and OApp ${oapp} to ${uln}`,
         }
     }
@@ -334,7 +369,7 @@ export class EndpointV2 extends OmniSDK implements IEndpointV2 {
 
                 const instruction = await this.program.setOappConfig(
                     this.connection,
-                    this.userAccount,
+                    await this.safeGetDelegate(oapp),
                     new PublicKey(oapp),
                     new PublicKey(uln),
                     setConfigParam.eid,
@@ -569,7 +604,13 @@ export class EndpointV2 extends OmniSDK implements IEndpointV2 {
         this.logger.verbose(`Initializing OApp nonce for OApp ${oapp} and peer ${peer} on ${eidLabel}`)
 
         const instruction = await mapError(
-            () => this.program.initOAppNonce(this.userAccount, eid, new PublicKey(oapp), normalizePeer(peer, eid)),
+            async () =>
+                this.program.initOAppNonce(
+                    await this.safeGetDelegate(oapp),
+                    eid,
+                    new PublicKey(oapp),
+                    normalizePeer(peer, eid)
+                ),
             (error) =>
                 new Error(
                     `Failed to init nonce for ${this.label} for OApp ${oapp} and peer ${peer} on ${eidLabel}: ${error}`
@@ -615,7 +656,7 @@ export class EndpointV2 extends OmniSDK implements IEndpointV2 {
         this.logger.verbose(`Initializing OApp send library for OApp ${oapp} on ${eidLabel}`)
 
         const instruction = await mapError(
-            () => this.program.initSendLibrary(this.userAccount, new PublicKey(oapp), eid),
+            async () => this.program.initSendLibrary(await this.safeGetDelegate(oapp), new PublicKey(oapp), eid),
             (error) =>
                 new Error(`Failed to init send library for ${this.label} for OApp ${oapp} on ${eidLabel}: ${error}`)
         )
@@ -654,7 +695,7 @@ export class EndpointV2 extends OmniSDK implements IEndpointV2 {
         this.logger.verbose(`Initializing OApp receive library for OApp ${oapp} on ${eidLabel}`)
 
         const instruction = await mapError(
-            () => this.program.initReceiveLibrary(this.userAccount, new PublicKey(oapp), eid),
+            async () => this.program.initReceiveLibrary(await this.safeGetDelegate(oapp), new PublicKey(oapp), eid),
             (error) =>
                 new Error(`Failed to init receive library for ${this.label} for OApp ${oapp} on ${eidLabel}: ${error}`)
         )
@@ -698,7 +739,7 @@ export class EndpointV2 extends OmniSDK implements IEndpointV2 {
         this.logger.verbose(`Initializing OApp config library for OApp ${oapp} on ${eidLabel}`)
 
         const instruction = await mapError(
-            () => {
+            async () => {
                 const libPublicKey = new PublicKey(lib)
                 const msgLibInterface = libPublicKey.equals(SimpleMessageLibProgram.PROGRAM_ID)
                     ? (this.logger.debug(`Using SimpleMessageLib at ${libPublicKey} to initialize OApp config`),
@@ -707,7 +748,7 @@ export class EndpointV2 extends OmniSDK implements IEndpointV2 {
                       new UlnProgram.Uln(libPublicKey))
 
                 return this.program.initOAppConfig(
-                    this.userAccount,
+                    await this.safeGetDelegate(oapp),
                     msgLibInterface,
                     this.userAccount,
                     new PublicKey(oapp),