+++ examples/oft-solana/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @layerzerolabs/oft-solana-example
 
+## 0.4.6
+
+### Patch Changes
+
+- d1d51ef: Bump ua-devtools-evm-hardhat dependency to 6.0.6+
+
+## 0.4.5
+
+### Patch Changes
+
+- 1d2abff: new SDK methods, tests in devtools-ton, upgraded lz-definitions
+
 ## 0.4.4
 
 ### Patch Changes