+++ packages/ua-devtools-solana/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @layerzerolabs/ua-devtools-solana
 
+## 4.0.5
+
+### Patch Changes
+
+- 51c4086: Fix setSendConfig and setReceiveConfig for squads multisig
+- Updated dependencies [51c4086]
+  - @layerzerolabs/protocol-devtools-solana@4.0.5
+
+## 4.0.4
+
+### Patch Changes
+
+- 1d2abff: new SDK methods, tests in devtools-ton, upgraded lz-definitions
+- Updated dependencies [1d2abff]
+  - @layerzerolabs/protocol-devtools-solana@4.0.3
+  - @layerzerolabs/protocol-devtools@1.1.2
+  - @layerzerolabs/devtools-solana@1.0.4
+  - @layerzerolabs/ua-devtools@3.0.2
+  - @layerzerolabs/devtools@0.4.3
+
 ## 4.0.3
 
 ### Patch Changes