+++ tests/ua-devtools-evm-hardhat-test/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @layerzerolabs/ua-devtools-evm-hardhat-test
 
+## 0.7.4
+
+### Patch Changes
+
+- d1d51ef: Bump ua-devtools-evm-hardhat dependency to 6.0.6+
+
+## 0.7.3
+
+### Patch Changes
+
+- 1d2abff: new SDK methods, tests in devtools-ton, upgraded lz-definitions
+
 ## 0.7.2
 
 ### Patch Changes